requirements.txt
django==4.1.3
requests==2.28.1
Pillow==9.3.0
opencv-python==4.6.0.66
numpy==1.23.4
urllib3==1.26.12